+++ examples/processTransaction.ts
@@ -26,7 +26,6 @@ async function main() {
     });
 
     const eventAdapter = sskts.adapter.event(connection);
-    const personAdapter = sskts.adapter.person(connection);
     const organizationAdapter = sskts.adapter.organization(connection);
     const transactionAdapter = sskts.adapter.transaction(connection);
     const transactionCountAdapter = sskts.adapter.transactionCount(redisClient);
@@ -48,12 +47,18 @@ async function main() {
         maxCountPerUnit: 120,
         scope: scope,
         clientUser: {
-            client: 'client',
-            state: 'state',
-            scopes: []
+            client_id: 'cq8em79ileeqlcp6q5kkd446n',
+            jti: 'cf7c2a05-5ab6-4fc6-a3c1-ec6c6fc9254f',
+            version: 2,
+            iat: 1502954230,
+            exp: 1502957830,
+            iss: 'iss',
+            scope: 'https://sskts-api-development.azurewebsites.net/transactions',
+            token_use: 'access',
+            sub: 'cq8em79ileeqlcp6q5kkd446n'
         },
         sellerId: movieTheaters[0].id
-    })(personAdapter, organizationAdapter, transactionAdapter, transactionCountAdapter);
+    })(organizationAdapter, transactionAdapter, transactionCountAdapter);
     if (transactionOption.isEmpty) {
         throw new Error('no ready transaction');
     }
@@ -223,7 +228,7 @@ async function main() {
         telephone: telephone,
         email: process.env.SSKTS_DEVELOPER_EMAIL
     };
-    await sskts.service.transaction.placeOrder.setAgentProfile(transactionId, profile)(personAdapter, transactionAdapter);
+    await sskts.service.transaction.placeOrder.setAgentProfile(transactionId, profile)(transactionAdapter);
     debug('agent profile updated.');
 
     // 取引成立