+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 注文作成取引に対するタスクに渡すデータを取引オブジェクトに変更。
 - indexモジュールを再構成。
 - update package [@motionpicture/gmo-service](https://www.npmjs.com/package/@motionpicture/gmo-service)
+- Amazon Cognitoでの会員管理に対応
 
 ### Deprecated